Move VideoPlayer component to better folder structure
