+++ src/components/VideoPlayer/VideoPlayer.scss
@@ -1,4 +1,4 @@
-@import "../scss/input";
+@import "../../scss/input";
 
 .VideoPlayer {
   width: min(90vw, 800px);
